Parse Uint32 instead of long long to get rid of C99 LLONG_* defines, show also signed value in DebubUI when highest bit is set.

+++ src/debug/evaluate.h
@@ -10,6 +10,6 @@
 
 extern bool Eval_Number(const char *value, Uint32 *number);
 extern int Eval_Range(char *str, Uint32 *lower, Uint32 *upper);
-extern const char* Eval_Expression(const char *expression, long long *result, int *offset);
+extern const char* Eval_Expression(const char *expression, Uint32 *result, int *offset);
 
 #endif
